+++ libs/base/src/math/CSparse/cs_qr.c
@@ -10,7 +10,7 @@ csn *cs_qr (const cs *A, const css *S)
     if (!CS_CSC (A) || !S) return (NULL) ;
     m = A->m ; n = A->n ; Ap = A->p ; Ai = A->i ; Ax = A->x ;
     q = S->q ; parent = S->parent ; pinv = S->pinv ; m2 = S->m2 ;
-    vnz = S->lnz ; rnz = S->unz ; leftmost = S->leftmost ;
+    vnz = (int)S->lnz ; rnz = (int)S->unz ; leftmost = S->leftmost ;
     w = cs_malloc (m2+n, sizeof (int)) ;            /* get int workspace */
     x = cs_malloc (m2, sizeof (double)) ;           /* get double workspace */
     N = cs_calloc (1, sizeof (csn)) ;               /* allocate result */
